optimize logger custom configuration
The previous logger custom configuration code was very redundant and
difficult to test. By refactoring the code of DefaultSdider and
LogConfiguration, it is now easy to test logger custom configuration.
After the refactoring, unnecessary LogAction classes were deleted, and
the new API is easier to understand than the previous ones.