improve implementation of global request config
Using static variables to implement global configuration is a bad
design, which will make the relevant code scattered and difficult
to understand. This commit added a RequestFactory interface to
determine how to create and initialize a Request. The default
implementation DefaultRequestFactory will initialize the request
with the global configuration while creating a request.
The DefaultRequestFactory is not thread-safe, which may leave hidden
dangers.